pyproject.toml
[project]
name = "kenya-tourism-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.2.10",
    "mistralai>=1.12.4",
    "pandas>=2.3.3",
    "pgvector>=0.4.2",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
]
